refactor: externalize websocket info component
